backend lab 4 sprint 1

+++ implementacao/backend/src/main/java/com/sistema_de_moeda_estudantil/sistema_de_moeda_estudantil/DTO/Aluno/AlunoCreate.java
@@ -26,7 +26,7 @@ public class AlunoCreate {
     private String endereco;
     private String curso;
 
-    private Integer instituicaoId;
+    private Long instituicaoId;
 }
