+++ implementacao/backend/src/main/java/com/sistema_de_moeda_estudantil/sistema_de_moeda_estudantil/DTO/Vantagem/VantagemCreate.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class VantagemCreate {
     private String descricao;
     private String foto;
-    private int custoMoedas;
-    private int quantidadeDisponivel;
+    private Long custoMoedas;
+    private Long quantidadeDisponivel;
 }
