+++ implementacao/backend/src/main/java/com/sistema_de_moeda_estudantil/sistema_de_moeda_estudantil/DTO/Vantagem/VantagemDTO.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class VantagemDTO {
-    private int id;
+    private Long id;
     private String descricao;
     private String foto;
-    private int custoMoedas;
-    private int quantidadeDisponivel;
+    private Long custoMoedas;
+    private Long quantidadeDisponivel;
 }
