+++ implementacao/backend/src/main/java/com/sistema_de_moeda_estudantil/sistema_de_moeda_estudantil/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sistema_de_moeda_estudantil.sistema_de_moeda_estudantil.entity.Usuario;
 
-public interface UserRepository extends CrudRepository<Usuario, Integer> {
+public interface UserRepository extends CrudRepository<Usuario, Long> {
     Optional<Usuario> findByEmail(String email);
   }
